MS44 lists moved to pages folders
